refactor(rapport): split bin into lib + main shim
Move all CLI logic into `rapport::run`, leaving `main.rs` a 5-line
shim that forwards `env::args().skip(1)`. Sets up the entry point
that future tests will call directly with fakes.
